requirements_dev.txt
pyyaml
Pillow